+++ lrvTrackDebug.cpp
@@ -33,10 +33,10 @@ void verbosePrint(std::string &toPrint)
     }
 }
 
-std::string printUIMap(std::map<unsigned int, unsigned int> &A)
+std::string printUIMap(std::map<size_t, size_t> &A)
 {
   std::stringstream F;
-  std::map<unsigned int, unsigned int>::iterator Ait=A.begin();
+  std::map<size_t, size_t>::iterator Ait=A.begin();
   F << "[ ";
   while(Ait!=A.end())
   {
